- scoped usage of  functions from applicant_review_processor.py - renamed applicant_review_processor function names - moved applicant_review_processor unit tests to new file
